add discord league to dashboard
